Move ExperimentList to ProjectInformationSupport and adapt view and routing
