+++ vaadinfrontend/frontend/themes/datamanager/components/main.css
@@ -1,14 +1,14 @@
 .main {
   display: grid;
   gap: 1rem 1px;
+  height: 100%;
 }
 
 .main.experiment {
-  grid-template-columns: minmax(min-content, 80%) minmax(min-content, 20%);
   grid-template-rows: minmax(130px, auto) 1fr;
   grid-template-areas:
-    "navbar support"
-    "content support";
+    "navbar"
+    "content";
   height: 100%;
 }
 
@@ -21,7 +21,8 @@
 }
 
 .main.project {
-  grid-template-columns: minmax(min-content, 80%) minmax(min-content, 20%);
+  grid-template-columns: minmax(min-content, 80%) minmax(min-content, 25%);
+  grid-template-rows: minmax(10px, 100%);
   grid-template-areas:
     "content support";
   height: 100%;
@@ -34,6 +35,7 @@
 
 .main.project .support {
   grid-area: support;
+  height: 100%;
 }
 
 .main.sample {
@@ -60,8 +62,7 @@
     grid-template-columns: minmax(1200px, 1fr);
     grid-template-areas:
         "navbar"
-    "content"
-    "support";
+    "content";
     grid-auto-rows: auto;
   }
