+++ vaadinfrontend/src/main/java/life/qbic/datamanager/views/projects/project/experiments/ExperimentContentComponent.java
@@ -1,16 +1,11 @@
 package life.qbic.datamanager.views.projects.project.experiments;
 
-import com.vaadin.flow.component.ComponentEvent;
-import com.vaadin.flow.component.ComponentEventListener;
 import com.vaadin.flow.component.html.Div;
 import com.vaadin.flow.spring.annotation.SpringComponent;
 import com.vaadin.flow.spring.annotation.UIScope;
 import java.io.Serial;
 import life.qbic.datamanager.views.Context;
-import life.qbic.datamanager.views.general.Disclaimer;
 import life.qbic.datamanager.views.projects.project.experiments.experiment.ExperimentDetailsComponent;
-import life.qbic.datamanager.views.projects.project.experiments.experiment.ExperimentDetailsComponent.ExperimentNameChangedEvent;
-import life.qbic.projectmanagement.domain.project.experiment.Experiment;
 import org.springframework.beans.factory.annotation.Autowired;
 
 /**
@@ -26,24 +21,13 @@ import org.springframework.beans.factory.annotation.Autowired;
 @SpringComponent
 @UIScope
 public class ExperimentContentComponent extends Div {
-
   @Serial
   private static final long serialVersionUID = 464171225772721108L;
   private final ExperimentDetailsComponent experimentDetailsComponent;
-  private final Disclaimer noExperimentDisclaimer;
 
   public ExperimentContentComponent(
       @Autowired ExperimentDetailsComponent experimentDetailsComponent) {
     this.experimentDetailsComponent = experimentDetailsComponent;
-    this.noExperimentDisclaimer = createNoExperimentDisclaimer();
-  }
-
-  private Disclaimer createNoExperimentDisclaimer() {
-    var disclaimer = Disclaimer.createWithTitle("Add an experiment",
-        "Get started by adding an experiment", "Add experiment");
-    disclaimer.addDisclaimerConfirmedListener(confirmedEvent -> fireEvent(
-        new AddExperimentClickEvent(this, confirmedEvent.isFromClient())));
-    return disclaimer;
   }
 
   /**
@@ -56,41 +40,7 @@ public class ExperimentContentComponent extends Div {
         experimentId -> {
           experimentDetailsComponent.setContext(context);
           this.add(experimentDetailsComponent);
-          this.remove(noExperimentDisclaimer);
-        }, () -> {
-          this.add(noExperimentDisclaimer);
-          this.remove(experimentDetailsComponent);
-        }
+        }, () -> this.remove(experimentDetailsComponent)
     );
   }
-
-  /**
-   * Propagates the listener which will retrieve notification if a an {@link Experiment} was edited
-   * in the {@link ExperimentDetailsComponent} within this container
-   */
-  public void addExperimentNameChangedListener(
-      ComponentEventListener<ExperimentNameChangedEvent> experimentEditListener) {
-    experimentDetailsComponent.addExperimentNameChangedListener(experimentEditListener);
-  }
-
-  public void addExperimentAddButtonClickEventListener(
-      ComponentEventListener<AddExperimentClickEvent> listener) {
-    addListener(AddExperimentClickEvent.class, listener);
-  }
-
-  public static class AddExperimentClickEvent extends ComponentEvent<ExperimentContentComponent> {
-
-    /**
-     * Creates a new event using the given source and indicator whether the event originated from
-     * the client side or the server side.
-     *
-     * @param source     the source component
-     * @param fromClient <code>true</code> if the event originated from the client
-     *                   side, <code>false</code> otherwise
-     */
-    public AddExperimentClickEvent(ExperimentContentComponent source, boolean fromClient) {
-      super(source, fromClient);
-    }
-  }
-
 }
